[xy] Create a new logger file.
