+++ mage_ai/data_cleaner/data_cleaner.py
@@ -2,7 +2,8 @@ from mage_ai.data_cleaner import column_type_detector
 from mage_ai.data_cleaner.analysis.calculator import AnalysisCalculator
 from mage_ai.data_cleaner.pipelines.base import BasePipeline
 from mage_ai.data_cleaner.shared.hash import merge_dict
-from mage_ai.data_cleaner.shared.utils import timer, clean_series, clean_dataframe
+from mage_ai.data_cleaner.shared.logger import timer
+from mage_ai.data_cleaner.shared.utils import clean_series, clean_dataframe
 from mage_ai.data_cleaner.statistics.calculator import StatisticsCalculator
