+++ mage_ai/data_cleaner/statistics/calculator.py
@@ -1,6 +1,6 @@
 from mage_ai.data_cleaner.shared.hash import merge_dict
 from mage_ai.data_cleaner.shared.multi import run_parallel
-from mage_ai.data_cleaner.shared.utils import timer
+from mage_ai.data_cleaner.shared.logger import timer
 from mage_ai.data_cleaner.column_type_detector import (
     DATETIME,
     NUMBER,
